Fetching of games from API. Styling
