+++ package.json
@@ -22,7 +22,7 @@
     "babel-plugin-macros": "^3.1.0",
     "eslint-plugin-react-native": "^4.0.0",
     "expo": "~47.0.12",
-    "expo-av": "~13.0.2",
+    "expo-av": "~13.0.3",
     "expo-image-picker": "~14.0.2",
     "expo-status-bar": "~1.4.2",
     "prop-types": "^15.8.1",
@@ -31,9 +31,10 @@
     "react-hook-form": "^7.43.1",
     "react-native": "0.70.5",
     "react-native-dropdown-select-list": "^2.0.4",
-    "react-native-gesture-handler": "^2.9.0",
+    "react-native-gesture-handler": "~2.8.0",
     "react-native-paper": "^5.2.0",
-    "react-native-reanimated": "2.9.1",
+    "react-native-reanimated": "~2.12.0",
+    "react-native-searchable-dropdown": "^1.1.3",
     "react-native-svg": "13.4.0",
     "react-native-svg-transformer": "^1.0.0",
     "react-native-toast-message": "^2.1.5",
